+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+WIP
+- add `add_limit_order`, `remove_limit_order` tx types
+
 ## 0.21.0 - 2021.04.09
 - add v2.0 tx types
 - add `payloadLength` param to `getFeeValue`